- changed Push/Pull subscribe options validation fix NPE - removed ignoring of negative values during SimpleJson toString b/c why plus it didn't account for long values representing nanos - split MsgHeader to own file and extracted construction of headers in preparation to handle Status object as a separate object
